requirements.txt
# WeTechForU AI Marketing Platform Dependencies

# Core Flask
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Security
Werkzeug==2.3.7
bcrypt==4.0.1

# HTTP Requests
requests==2.31.0
urllib3==2.0.4

# Environment
python-dotenv==1.0.0

# Email
smtplib2==0.2.1

# Data Processing
pandas==2.1.1
numpy==1.24.3

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.0

# API Integration
facebook-sdk==3.1.0
google-ads==22.1.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
stripe==6.7.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# Development
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0

# Production
gunicorn==21.2.0
whitenoise==6.5.0
